Simulation quality assessment

Evaluates the simulated water budget with the average KGE.

Usage

compute_simulation_quality_assessment(
  obj,
  water_budget,
  rcn_gauging,
  observed_flow,
  alpha_lyne_hollick,
  period = NULL
)

Arguments

obj

The HydroBudget object with calibration parameters and column names mappings.

water_budget

The computed water budget. Input can be a data.frame/data.table or a path to a data file.

rcn_gauging

The table with the list of RCN cells located in each gauging station watershed. Input can be a data.frame/data.table or a path to a data file.

observed_flow

The flow rates in mm/day. Input can be a data.frame/data.table or a path to a data file.

alpha_lyne_hollick

The Lyne and Hollick filter. Input can be a data.frame/data.table or a path to a data file.

period

The start and end years. If not provided, the start/end years will be extracted from the water budget data.

Details

The columns of the water budget data set input are:

year

month

vi

t_mean

runoff

pet

aet

gwr

runoff_2

delta_reservoir

rcn_id

The columns of the observed flow data set input are:

year

month

day

one column per station (named by the station ID), the flow rates in mm/day

The columns of the RCN gauging stations data set input are:

rcn_id, the cell ID

station_id, the station ID

The columns of the Lyne and Hollick filter data set input are:

station_id, the station ID

alpha

Examples

if (FALSE) {
# Use input example files provided by the package
base_url <- "https://github.com/gwrecharge/rechaRge-book/raw/main/examples/input/"
input_rcn_gauging <- paste0(base_url, "rcn_gauging.csv.gz")
input_observed_flow <- paste0(base_url, "observed_flow.csv.gz")
input_alpha_lyne_hollick <- paste0(base_url, "alpha_lyne_hollick.csv.gz")

# Calibration parameters
HB <- rechaRge::new_hydrobudget(
  T_m = 2.1, # melting temperature (°C)
  C_m = 6.2, # melting coefficient (mm/°C/d)
  TT_F = -17.6, # Threshold temperature for soil frost (°C)
  F_T = 16.4, # Freezing time (d)
  t_API = 3.9, # Antecedent precipitation index time (d)
  f_runoff = 0.63, # Runoff factor (-)
  sw_m = 431, # Maximum soil water content (mm)
  f_inf = 0.07 # infiltration factor (-)
)

# ... compute the water budget ...

result <- compute_simulation_quality_assessment(
  HB,
  water_budget = water_budget,
  rcn_gauging = input_rcn_gauging,
  observed_flow = input_observed_flow,
  alpha_lyne_hollick = input_alpha_lyne_hollick,
  period = simul_period
)
}